Added plugin architecture and moved protect_files into plguin
